remaining logic completed — except livestream functionality and other minors

+++ server/app/Http/Controllers/Api/V1/PostController.php
@@ -9,12 +9,14 @@ use App\Models\Transaction;
 use App\Models\Notification;
 // use Illuminate\Http\Request;
 use App\Models\Subscription;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Models\Internaltransaction;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\PostResource;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
+use App\Models\Block;
 
 // use CloudinaryLabs\CloudinaryLaravel\Facades\Cloudinary;
 
@@ -22,7 +24,7 @@ class PostController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth:api', ['except' => ['index', 'featuredPosts']]);
+        $this->middleware('auth:api', ['except' => ['featuredPosts']]);
     }
 
     /**
@@ -30,19 +32,99 @@ class PostController extends Controller
      */
     public function index()
     {
-        // $posts = Post::withTrashed()->latest()->paginate();
-        $posts = Post::with([
-            'user',
-            'comments',
-            'comments.user',
-            'likes',
-            'likes.user',
-            'bookmarks',
-            'bookmarks.user',
-            'bookmarks.post',
-            'bookmarks.post.comments',
-            'bookmarks.post.likes',
-        ])->latest()->paginate();
+        // $blocking = Block::where('blocked_id', auth()->user()->id)->first();
+
+        // $subscription = Subscription::where('subscriber_id', auth()->user()->id)->first();
+
+        // if ($blocking && $subscription) {
+        //     $posts = Post::with([
+        //         'user',
+        //         'comments',
+        //         'comments.user',
+        //         'likes',
+        //         'likes.user',
+        //         'bookmarks',
+        //         'bookmarks.user',
+        //         'bookmarks.post',
+        //         'bookmarks.post.comments',
+        //         'bookmarks.post.likes',
+        //     ])
+        //         ->where([
+        //             'user_id' => auth()->user()->id,
+        //             'featured' => true,
+        //             'user_id', '!=', $blocking?->blocker_id,
+        //             'user_id', '=', $subscription?->subscribed_id
+        //         ])->latest()->paginate();
+
+        //     return PostResource::collection($posts);
+        // }
+
+        // if ($blocking) {
+        //     $posts = Post::with([
+        //         'user',
+        //         'comments',
+        //         'comments.user',
+        //         'likes',
+        //         'likes.user',
+        //         'bookmarks',
+        //         'bookmarks.user',
+        //         'bookmarks.post',
+        //         'bookmarks.post.comments',
+        //         'bookmarks.post.likes',
+        //     ])
+        //         ->where([
+        //             'user_id' => auth()->user()->id,
+        //             'featured' => true,
+        //             'user_id', '!=', $blocking?->blocker_id
+        //         ])->latest()->paginate();
+
+        //     return PostResource::collection($posts);
+        // }
+
+        // if ($subscription) {
+        //     $posts = Post::with([
+        //         'user',
+        //         'comments',
+        //         'comments.user',
+        //         'likes',
+        //         'likes.user',
+        //         'bookmarks',
+        //         'bookmarks.user',
+        //         'bookmarks.post',
+        //         'bookmarks.post.comments',
+        //         'bookmarks.post.likes',
+        //     ])
+        //         ->where([
+        //             'user_id' => auth()->user()->id,
+        //             'featured' => true,
+        //             'user_id', '=', $subscription?->subscribed_id
+        //         ])->latest()->paginate();
+
+        //     return PostResource::collection($posts);
+        // }
+
+        // if (!$blocking && !$subscription) {
+        //     $posts = Post::with([
+        //         'user',
+        //         'comments',
+        //         'comments.user',
+        //         'likes',
+        //         'likes.user',
+        //         'bookmarks',
+        //         'bookmarks.user',
+        //         'bookmarks.post',
+        //         'bookmarks.post.comments',
+        //         'bookmarks.post.likes',
+        //     ])
+        //         ->where([
+        //             ['user_id' => auth()->user()->id] ||
+        //                 ['featured' => true],
+        //         ])->latest()->paginate();
+
+        //     return PostResource::collection($posts);
+        // }
+
+        $posts = Post::latest()->paginate();
 
         return PostResource::collection($posts);
     }
@@ -66,12 +148,11 @@ class PostController extends Controller
             $post['image_url'] = $path;
         }
 
-        if ($request->file('video_url')) {
+        if ($request->hasFile('video_url')) {
             $path = $validated['video_url']->store('videos/posts');
             $post['video_url'] = $path;
         }
 
-        $post['user_id'] = $validated['user_id'];
         $post['body'] = $validated['body'];
 
         $post->save();
@@ -199,7 +280,26 @@ class PostController extends Controller
             abort(403);
         }
 
-        $post->update($request->validated());
+        $validated = $request->validated();
+
+        if ($request->file('image_url')) {
+            // Upload an Image File to Cloudinary with One line of Code
+            // $uploadedFileUrl = Cloudinary::upload($validated['image_url'])->getRealPath())->getSecurePath();
+            // // $uploadedFileUrl = Cloudinary::upload($request->file('image_url')->getRealPath())->getSecurePath();
+            // $post['image_url'] = $uploadedFileUrl;
+
+            $path = $validated['image_url']->store('images/posts');
+            $post['image_url'] = $path;
+        }
+
+        if ($request->file('video_url')) {
+            $path = $validated['video_url']->store('videos/posts');
+            $post['video_url'] = $path;
+        }
+
+        $post['body'] = $validated['body'];
+
+        $post->update();
 
         return new PostResource($post);
     }
@@ -207,8 +307,12 @@ class PostController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Post $post)
+    public function destroy(Request $request, Post $post)
     {
+        if ($request->user()->cannot('delete', $post)) {
+            abort(403);
+        }
+
         $post->delete();
     }
 
@@ -236,8 +340,12 @@ class PostController extends Controller
     /**
      * Repost a post.
      */
-    public function repost(StorePostRequest $request)
+    public function repost(StorePostRequest $request, Post $post)
     {
+        if ($request->user()->cannot('repost', $post)) {
+            abort(403);
+        }
+
         $validated = $request->validated();
 
         $re_check_if_post_exists = Post::where('id', $validated['repost_original_id'])->first();
